% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiplot.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{plotlist}{Una lista de los gráficos a graficar en el arreglo.}

\item{cols}{Número de columnas del arreglo.}

\item{layout}{El layout a utilizar.}
}
\value{
Un gráfico en forma de arreglo
}
\description{
Función que realiza un arreglo de varias gráficas.
}
