% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPS.R
\name{plotMPS}
\alias{plotMPS}
\title{Plot an object of class MPS}
\usage{
plotMPS(MPSObject, qLoss, qDistance)
}
\arguments{
\item{MPSObject}{Un objeto de clase MPS (FasMPS o EvalMPS).}

\item{qLoss}{Cuantil para la pérdida.}

\item{qDistance}{Cuantil para la distancia.}
}
\value{
Un gráfico de la salida de MPS.
}
\description{
Función que gráfica las salidas de la función MPS.
}
