% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPS2.R
\name{plotMPS2}
\alias{plotMPS2}
\title{Plot an object of class MPS}
\usage{
plotMPS2(MPSObject, qLoss, qSimilarity)
}
\arguments{
\item{MPSObject}{Un objeto de clase MPS (ApproxMPS).}

\item{qLoss}{Cuantil para la pérdida.}

\item{qSimilarity}{Cuantil para la similaridad.}
}
\value{
Un gráfico de la salida de MPS.
}
\description{
Función que gráfica las salidas de la función ApproxMPS.
}
