% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPSFromG.R
\name{plotMPSFromG}
\alias{plotMPSFromG}
\title{Plot an object of class MPS}
\usage{
plotMPSFromG(MPSObject, Sim, qLoss, qSimI, qSimS)
}
\arguments{
\item{MPSObject}{Un objeto de clase MPS (ApproxMPS).}

\item{Sim}{vector de similaridades promedio para cada línea candidata.}

\item{qLoss}{Cuantil para la pérdida.}

\item{qSimI}{Cuantil inferior para la similaridad.}

\item{qSimS}{Cuantil superior para la similaridad.}
}
\value{
Un gráfico de la salida de MPS.
}
\description{
Función que gráfica las salidas de la función ApproxMPS.
}
