% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPSFromPed.R
\name{plotMPSFromPed}
\alias{plotMPSFromPed}
\title{Plot an object of class MPS}
\usage{
plotMPSFromPed(MPSObject, Sim, qLoss, qSimilarity)
}
\arguments{
\item{MPSObject}{Un objeto de clase MPS (ApproxMPS).}

\item{Sim}{vector de similaridades promedio para cada línea candidata.}

\item{qLoss}{Cuantil para la pérdida.}

\item{qSimilarity}{Cuantil para la similaridad.}
}
\value{
Un gráfico de la salida de MPS.
}
\description{
Función que gráfica las salidas de la función ApproxMPS.
}
