% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPSggplot.R
\name{plotMPSggplot}
\alias{plotMPSggplot}
\title{Plot an object of class MPS}
\usage{
plotMPSggplot(MPSObject, Sim, qLoss, qSimilarity, B0, bandera = FALSE)
}
\arguments{
\item{MPSObject}{Un objeto de clase MPS (ApproxMPS).}

\item{Sim}{vector de similaridades promedio para cada línea candidata.}

\item{qLoss}{Cuantil para la pérdida.}

\item{qSimilarity}{Cuantil para la similaridad.}

\item{B0}{(vector) with the overall mean of each trait of length equal to number of traits (\eqn{t}).}

\item{bandera}{Un indicador de si B0 son ceros.}
}
\value{
Un gráfico de la salida de MPS.
}
\description{
Función que gráfica las salidas de la función ApproxMPS.
}
