% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotMPS.R
\name{plotMPS}
\alias{plotMPS}
\title{Plot an object of class MPS}
\usage{
plotMPS(MPSObject, Dist, qLoss, qDistance, B0)
}
\arguments{
\item{MPSObject}{Un objeto de clase MPS (FasMPS o EvalMPS).}

\item{Dist}{vector de distancias promedio de cada línea candidata.}

\item{qLoss}{Cuantil para la pérdida.}

\item{qDistance}{Cuantil para la distancia.}

\item{B0}{(vector) with the overall mean of each trait of length equal to number of traits (\eqn{t}).}
}
\value{
Un gráfico de la salida de MPS.
}
\description{
Función que gráfica las salidas de la función MPS.
}
